Fixed: modify routes order of operations
ModifyRoutes.js may kick off multiple REST calls to posts.js to modify directories and files in Routes. async/await syntax was getting in the way. If more than one route was being modified, the first would fire and complete successfully. But the rest would be ignored because live-server would trigger a page refresh, killing the running script and stopping the rest of the calls from firing.

Removing the await keyword in the body of the for loop beginning on line 230 is all that was needed. I'm not sure if this is enough. They all kick off at the same time now. But if one of the calls completes before the rest, it may trigger live-server to reload. This may not matter. Once the calls are kicked off, I don't think they have to return a value for posts.js to complete it's tasks.